several player movement improvements
also decided to replace the player instance in the game scene. Trying to remove the overrides on it in the game scene and actually putting them in the prefab ended up breaking that player instance, so i just deleted and dragged it back in and it all works once again.

in the future please remember to add stuff into the prefab itself... i dont want to see entire scripts that should belong on the player only be on the player instance in the game scene. which i found 2 of.